Update 2022-07-20-wei22a.md
Dear Editor, 
I've tried everything I can to deal with the '&' special letter in the tex_title, still, report an error in the BibTeX citation. I have to report that the special letter '&' in network transmission perhaps really wouldn't work for git. the protocol automatically adds the 'amp' ';' after it. That could be a potential problem with using Git for papers.

So, I will have to give up my original '&' in my title, unfortunately. I now replace it with the word 'and'. Sorry to keep pulling requests. I really would like to fix our citation right, because the paper is the effort of our whole team. Hope you could understand:)
I believe this is the last edit. 

Many thanks!

Best regards,
Xiaoxi